Centralized functions used in multiple modules in the potiron module
